Adding React Router v4
