+++ config/webpack.config.prod.js
@@ -101,6 +101,11 @@ module.exports = {
     publicPath: publicPath
   },
   resolve: {
+    modules: [
+      paths.appSrc,
+      'node_modules',
+      paths.appContracts,
+    ], 
     // These are the reasonable defaults supported by the Node ecosystem.
     // We also include JSX as a common component filename extension to support
     // some tools, although we do not recommend using it, see:
