+++ package.json
@@ -62,7 +62,8 @@
     "final-form": "^4.2.1",
     "material-ui": "^1.0.0-beta.35",
     "material-ui-icons": "^1.0.0-beta.35",
-    "react-final-form": "^3.1.2"
+    "react-final-form": "^3.1.2",
+    "react-router-dom": "^4.2.2"
   },
   "jest": {
     "collectCoverageFrom": [
