+++ src/routes/welcome/components/App.scss
